python/requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
flask==3.0.0
python-dotenv==1.0.0
face-recognition==1.4.0
dlib==19.24.2
Pillow==10.0.0
requests==2.31.0
scipy==1.11.2
